Fix infinite loop on error conditions that are not caught
There are cases where nmcli returns other errors which are not handled by wifi-control. This can cause an infinite loop when "Waiting for connection attempt to settle" loop runs. I added a catch all which gets around this. An example of a way to reproduce these other errors is by not having permissions to create connections on the host system by the user that is running Node. Usually a user other than root.